Renamed planscore.api_states to planscore.get_states
